refactor(auth): extract the OAuth token-endpoint layer out of the runtime
The refresh and revoke endpoint helpers (the grant_type / URLSearchParams
protocol code) move out of create-oauth-app-auth.ts into the module that
already owns token-grant parsing, which is renamed oauth-token-response.ts
-> oauth-token-endpoints.ts to reflect that it now owns both the request
and response halves of the token endpoints. The runtime file drops ~70
lines and is left as the auth state machine and transports.

AuthFetch moves to auth-contract.ts alongside AuthClient/AuthState, where
the transport contract belongs. Five files were already reaching into the
720-line runtime module just for this type; they now import it from the
contract, and the relocation keeps the new dependency acyclic (the
endpoint module no longer needs a back-edge into the runtime).
