+++ packages/auth/src/index.ts
@@ -1,4 +1,4 @@
-export type { AuthClient, AuthState } from './auth-contract.js';
+export type { AuthClient, AuthFetch, AuthState } from './auth-contract.js';
 export * from './auth-errors.js';
 export {
 	ApiSessionResponse,
@@ -7,7 +7,6 @@ export {
 	UserId,
 } from './auth-types.js';
 export {
-	type AuthFetch,
 	type CreateOAuthAppAuthConfig,
 	createOAuthAppAuth,
 } from './create-oauth-app-auth.js';
