+++ packages/auth/src/node/oob-launcher.test.ts
@@ -18,7 +18,7 @@ import { expect, test } from 'bun:test';
 import { EPICENTER_OAUTH_SCOPE } from '@epicenter/constants/oauth';
 import type { Result } from 'wellcrafted/result';
 import { expectErr, expectOk } from 'wellcrafted/testing';
-import type { AuthFetch } from '../create-oauth-app-auth.js';
+import type { AuthFetch } from '../auth-contract.js';
 import type { OAuthLaunchResult } from '../oauth-launchers/contract.js';
 import { createOobOAuthLauncher } from './oob-launcher.js';
